# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "LandlordIn": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "type": "string"
          },
          "agreement_expiry_date": {
            "format": "date",
            "type": [
              "string",
              "null"
            ]
          },
          "agreement_start_date": {
            "format": "date",
            "type": [
              "string",
              "null"
            ]
          },
          "bank_name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "contact_person": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "iban": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "monthly_rent": {
            "type": [
              "number",
              "null"
            ]
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          },
          "name_ar": {
            "maxLength": 160,
            "type": [
              "string",
              "null"
            ]
          },
          "qid_cr_expiry_date": {
            "format": "date",
            "type": [
              "string",
              "null"
            ]
          },
          "qid_cr_number": {
            "type": "string"
          },
          "remarks": {
            "type": "string"
          },
          "reminder_days_before_expiry": {
            "default": 90,
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "LandlordUpdateIn": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "type": [
              "string",
              "null"
            ]
          },
          "agreement_expiry_date": {
            "format": "date",
            "type": [
              "string",
              "null"
            ]
          },
          "agreement_start_date": {
            "format": "date",
            "type": [
              "string",
              "null"
            ]
          },
          "bank_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "contact_person": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "iban": {
            "type": [
              "string",
              "null"
            ]
          },
          "mobile": {
            "type": [
              "string",
              "null"
            ]
          },
          "monthly_rent": {
            "type": [
              "number",
              "null"
            ]
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          },
          "name_ar": {
            "maxLength": 160,
            "type": [
              "string",
              "null"
            ]
          },
          "qid_cr_expiry_date": {
            "format": "date",
            "type": [
              "string",
              "null"
            ]
          },
          "qid_cr_number": {
            "type": [
              "string",
              "null"
            ]
          },
          "remarks": {
            "type": [
              "string",
              "null"
            ]
          },
          "reminder_days_before_expiry": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "detail": {
            "properties": {
              "<location>": {
                "properties": {
                  "<field_name>": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  }
                },
                "type": "object"
              }
            },
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "title": "GreenTech Real Estate API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/landlords": {
      "get": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          }
        },
        "summary": "List Landlords"
      },
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LandlordIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Create Landlord"
      }
    },
    "/api/v1/landlords/{landlord_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "landlord_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Get Landlord"
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "landlord_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LandlordUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Update Landlord"
      }
    },
...
  },
...
}